added helpers.js
