+++ src/Dropdown.js
@@ -1,29 +1,9 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import classNames from 'classnames';
+import { isValidLabelOrValue, getOptionLabel, getOptionValue } from './helpers';
 
 const DEFAULT_PLACEHOLDER_STRING = 'Select...';
-
-export const isValidLabelOrValue = value => (
-  /string|boolean|number/.test(typeof value));
-
-export const getOptionLabel = (option, label = option) => {
-  if (isValidLabelOrValue(option.label))
-    label = option.label;
-  else if (isValidLabelOrValue(option.value))
-    label = option.value;
-
-  return label;
-};
-
-export const getOptionValue = (option, value = option) => {
-  if (isValidLabelOrValue(option.value))
-    value = option.value;
-  else if (isValidLabelOrValue(option.label))
-    value = option.label;
-
-  return value;
-};
 
 export const Option = props => {
   const { option, selected, onSelect, className } = props;
