+++ src/helpers.js
@@ -1,0 +1,20 @@
+export const isValidLabelOrValue = value => (
+  /string|boolean|number/.test(typeof value));
+
+export const getOptionLabel = (option, label = option) => {
+  if (isValidLabelOrValue(option.label))
+    label = option.label;
+  else if (isValidLabelOrValue(option.value))
+    label = option.value;
+
+  return label;
+};
+
+export const getOptionValue = (option, value = option) => {
+  if (isValidLabelOrValue(option.value))
+    value = option.value;
+  else if (isValidLabelOrValue(option.label))
+    value = option.label;
+
+  return value;
+};
